+++ maboss/__init__.py
@@ -2,8 +2,8 @@ from .network import *
 from .simulation import *
 from .result import *
 from .gsparser import load
-from .maboss_server import MaBoSSClient
-from .umbs import UpP_MaBoSS
+from .server import MaBoSSClient
+from .upp import UpdatePopulation
 from .ensemble import EnsembleResult
 
 from colomoto_jupyter import IN_IPYTHON
