--- maboss/umbs/__init__.py
@@ -1,1 +1,0 @@
-from .umbs import UpP_MaBoSS
