+++ maboss/upp/__init__.py
@@ -1,0 +1,1 @@
+from .upp import UpdatePopulation
